NeighborlyAPI/requirements.txt
azure-functions==1.2.1
pymongo==4.2.0
feedgen==0.9.0
werkzeug==0.16.0

